GAMutatorMulti: fix uniform distribution of mutators.
The previous implementation only gave an uniform distribution with <= 2
mutators; else, in a series of N mutators, only the first one had
probability 1/N, while the other ones had decreasing ones (e.g. the
second had 1/N - 1/N^2) except the very last, which was too likely to be
invoked.

Also changed the receiver of Mutate() to a pointer receiver, which is
more correct (the value receiver only happened to work because the
fields of GAMutatorMulti are all slices, i.e. references.